use controllers to be consistent in db
